Close parentless popups
A parentless popup is not visible anywhere, and is not even associated
to a QML engine. Attempting to execute transitions would crash.

Make sure to remove any event filters (when using a plain QML Window,
like with QML TestCase) when a popup is made parentless. This ensures
that if the global shared tooltip was shown on an item that was
destroyed, the tooltip won't keep interfering with the window where
it was last shown.

Change-Id: I87bf9f8dda2fbcc5d1b83d1b83a66a6b9e718eb6
Task-number: QTBUG-55347
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>