QQuickToolTip: fix the creation context of the shared tooltip
Don't create the shared tooltip in the parent item's context, because
that may be destroyed and invalidate all bindings in ToolTip.qml. The
QML engine is the effective parent of the shared tooltip, so use the
engine's root context for creation as well. We do not need to pass any
specific context to QQmlComponent::create(), because it will pick the
root context by default.

Change-Id: I624b53fa9ba02123398e76b5eccf1a8091826b37
Task-number: QTBUG-55347
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>