Init QQuickStyle's base URL to avoid creating temp QQmlEngine
Initialize the base URL from QtQuickControls2Plugin::registerTypes()
to avoid creating an unnecessary temporary instance of QQmlEngine. It
was used to list all known import paths, which were scanned for the
requested style folder. Now that we pass the base URL, the built-in
styles (a name without path to the style) can be looked up straight
from the base path.

Change-Id: I0d68d4c5bb90a4ab89660b5ce97f6433aeefc1ff
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>