Fix FileDialog::selectedNameFilter review findings
FileDialog::nameFilters contains an array of name filter strings that
combine the name and extensions. For example:
 - "Text files (*.txt)"
 - "HTML files (*.html *.htm)"

When dealing with multiple name filters, it is quite clumsy to use
a string type of 'selectedNameFilter' which is one of the above full
file name filter string values. Make it possible to read/write the
index of the selected name filter, and provide the filter's name and
extensions separately.

Change-Id: Ie416cc4ab3dcde93c10769b6f7ac44915307f194
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>