add QQuickEventPoint::sceneGrabPos property
Some handlers need to track total movement of the points from the time
that the gesture begins.  So far it seems that will always be the same
as the moment that the handler decides to grab; so it's enough to
store a copy of the current position when the grab occurs.
Also change setGrabber to call either setItemGrabber
or setPointerHandlerGrabber so that we only need to add
new features like this one in two places instead of three.

Change-Id: Ia944bf27302fc305a21957e8b02174909c024a90
Reviewed-by: Jan Arve Sæther <jan-arve.saether@theqtcompany.com>