Fix mouse area pressed after removal of pre-grab
Qt Quick used to actively "grab" the mouse before delivering events,
making the code more complex than necessary. Some items came to rely on
the behavior and now lack the ungrab event. This is triggered when
inside the delivery of a mouse event, some other item grabs the mouse.
If the mouse grab changed to an item that was not the target of the
original delivery, the grab must have been stolen, so call
mouseUngrabEvent in this case.

Test case by J-P Nurmi.

Task-number: QTBUG-55325
Change-Id: I2f0ac9d8aed1415662196070fb763f2752004d22
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>