Ensure windows renderloop doesn't stop existing, running, animations
If there is a normal C++ Qt animation running when the animation
driver is installed, and we don't have a visible QQuickWindow
at that time, the animation will await the QQuickWindow's vsync
to be ticked (which might never happen). So don't install
the driver until after we've registered started/stopped
(which will be called by install) and when m_vsyncDelta is
correctly initialized.

Change-Id: If36633ba78d8b16afa658ae5279372158a4f7924
Task-number: QTBUG-55521
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>