simplify QQuickPointerSingleHandler::wantsPointerEvent
Storing both the current point ID and the QQuickEventPoint pointer is
redundant; it was done because handlePointerEventImpl can then call
handleEventPoint() directly with that pointer.  But pointById isn't
so expensive.  Also, after 8a06075f48ad352acfc70111682e7decf8fedb33
if the handler is the grabber, the item will not be, so we don't need
to consider that case as before, where the item was the grabber on
behalf of the handler.

Change-Id: I3657ee46c823aa955e7bfd7714bbb4c5776f7d17
Reviewed-by: Jan Arve Sæther <jan-arve.saether@theqtcompany.com>