Fix conversion of QByteArray back to String in JavaScript
Commit 3b7e2a69f7eb8597c807de39b4de39721e9e2bd2 changed behavior so that
QByteArray is converted to the JS ArrayBuffer type, which is a better
fit than QVariant. However ArrayBuffer does not have a toString method
in the spec, and therefore any previous implicit toString conversion
such as when passing to JSON.parse() would fail. To restore
compatibility this patch adds a non-spec toString() that performs an
UTF-8 conversion, as it was done previously through QVariant's toString.

Task-number: QTBUG-55562
Change-Id: I096046954f7b29f7258deaa9ef5c8fa9292552ef
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>