Popup: respect dynamic dim & modal changes
What comes to the visual background dimming, popups did not respect
dynamic 'dim' or 'modal' property changes. This was half-intentional,
because it was thought an unrealistic use-case to toggle these while
the popup is visible. As a result, the background dimming was missing
if a popup was made visible in a declarative way, and either 'dim' or
'modal' was evaluated after the popup was already visible.

Task-number: QTBUG-54797
Change-Id: Icfb171e619540fef2d91908867b702790cde6c32
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>