Fix QQuickDrawerPrivate::handleMouseReleaseEvent()
QQuickDrawer operates on scene/window coordinates, se it must use
QMouseEvent::windowPos() instead of QMouseEvent::x() or y() which
return local coordinates. QMouseEvent::windowPos() is already used
in handleMousePressEvent() and handleMouseMoveEvent(). The delta
between the press and the release point is used to calculate the
velocity. The two points must be specified in the same coordinate
space to get correct velocity calculation.

Change-Id: Ic9ee904e0b2b022a58b674ee663d8963eb2417a6
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>