Fix style name casing
When QQuickStyle looks up the appropriate folder where the specified
style exists, it also fixes up the casing of the style name to match
what is on the file system. For example, "material" becomes "Material".

Since 54480f3, QtQuickControls2Plugin initializes the default style
path so QQuickStyle no longer has to create a temporary QQmlEngine to
be able to lookup styles in one of the import paths. It still needs
to fixup the casing of the style name to ensure that "-style material"
and QT_QUICK_CONTROLS_STYLE=material work.

Change-Id: If35b870bb335e024b94fb328a927a5550b51cba1
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>