Fix rotation in PinchHandler
If we want to allow rotation with more than 2 fingers, its not
straightforward to calculate the rotation angle, because we cannot anymore
calculate the angle between the two touchpoints.

The approach chosen was to calculate the angle between the centroid and
the touchpoints, and take the average between the angles.

Then, for the next event we calculated the new average of the angles.

However, this is not really reliable in some scenarios, suppose we have
these three angles:

0 120 240,   avg = 360/3 = 120

next, touch points rotate clockwise with 2 degrees, and we get these
angles:

358 118 238, avg = 714/3 = 238

So, just by rotating all fingers by 2 degrees, we got a jump by 118
degrees "in average".

Instead we need to track the angles of *all* touch points, and when the
next touch event is received we calculate how much the angle has changed
per touch point. We then take the average of those angles as the effective
"rotation" of the PinchHandler

Change-Id: I2bfdf80b886751177efe81bcc7b698af0d2938e3
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>