Fix tst_Drawer::dragMargin()
The width of the window in the test is 400px, and the
width of the drawer is 200px. The default drag margin
equals to QStyleHints::startDragDistance(), which reports
19px on Linux installed on a Chromebook Pixel 2.

The test attempts to drag from (double drag margin):

    2 x startDragDistance = 38px

to:

    0.25 x width of the drawer = 50px

12px does not exceed the start drag distance of 19px, so the drawer
is not dragged even though the test expects it to be dragged to the
logical position of 0.25. Increase the dragged distance a bit to
exceed the threshold.

Change-Id: Id469533b08c4221430446a2d995e36b4ebdc0258
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>