Remove superfluous QQuickOverlayPrivate::popups
First of all, the name was a bit misleading. It was a list of _visible_
popups. Anyway, it is unnecessary to maintain such list anymore. The
only use case for it was to resize the background dimming items of
visible popups when the window was resized. For that infrequent desktop
-only event, we can fetch the background dimming items from the list of
all popups.

Change-Id: Ie17db042719d4614189a78a2e5314136ca95df42
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>