QQuickDrawer: fix the internal transition state
When QQuickDrawer was visible all the time, it did not emit the
visibility related signals at all. Now the signals are emitted,
but aboutToShow() and visibleChanged() were emitted twice when
a drawer was manually dragged open. First time when the dragging
begins (the drawer becomes effectively visible), and second time
when the open transition begins after mouse/touch release. This
change ensures that the signals won't be emitted again when the
transition begins, in case they were already emitted.

Change-Id: I2a175c9e86a480d5cd23e306f41f0d85e2416f75
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>