Do not use manhattanLength for measuring distance
Since we use averageStartingDistance and averageTouchPointDistance to calculate
the scale factor for the PinchHandler, we got some unacceptable scale artifacts,
especially when rotating (for a 45 degree rotation, the manhattan length is
sqrt(2) times longer than the actual length). This is (probably) a bit slower,
but the number of points are usually quite low (range 2-5, where 2 is the most
common, so the sqrt should not be calculated very often.

Change-Id: Ibb3d4cc036a4acdd86a5c79c852ccf7df7ef4ebc
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>