QQuickTextInput: avoid emitting editingFinished twice
On platforms like iOS, hiding the virtual keyboard will force
the current focus object to lose focus. And when QQuickTextInput
loses focus, it will emit editingFinished.

To not emit the signal once more directly after hiding the
keyboard, we add an extra check in the key event handler
to verify that we still have focus.

Task-number: QTBUG-44038
Change-Id: Ib64f7028a9f16a172b3c38cab72f3599d560e331
Reviewed-by: Liang Qi <liang.qi@qt.io>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>