Doc: highlight examples on the index page
This looks nicer that an "Examples" section with a bullet list that
only contains a lonely "Qt Quick Controls 2 Examples" link. Manually
listing those two examples that we currently have in the 5.7 branch
together with an "All Examples" link may seem a bit silly for the time
being, but the idea is that later on we highlight a few examples
manually and the rest are accessible through the "All Examples" link.

Change-Id: I329d812b7d269de607fa141552e35bae76041176
Reviewed-by: Venugopal Shivashankar <Venugopal.Shivashankar@qt.io>