Fix crash when using Software Renderer on HiDPI QBackingStore
For some reason the QPaintDevice of a QBackingStore is not valid on
HiDPI QBackingStores until QBackingStore::begin() has been called.  I
suspect this is a bug in qtbase, but until this is fixed this change
will make sure that the paint device will always be valid, even in this
unusual situation.

Task-number: QTBUG-55875
Change-Id: I63137d0de9b68d29bfcf86d1807f576923875b5d
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>