Fix V4 on big-endian
We can't both invert offset position and inter-value positions of tag
and value.

This patch changes 32-bit big-endian to use the same order inside
the tag/value but just at different offsets. This also make it
compatible with how we use it with doubles.

This fixes value/tag reads on 32-bit big-endian and offsets on 64-bit.

Change-Id: I95cf792c29ac3f42a4018ce1f115193c143a0df0
Reviewed-by: Erik Verbruggen <erik.verbruggen@qt.io>