QQuickSwitch: fix event handling
Switch implemented custom event handling for the indicator, and used
QQuickAbstractButton's event handling for the background. This lead to
inconsistent signals depending on whether interacting with the handle
or the background. This change gets rid of the child mouse event filter
for the indicator and makes QQuickSwitch fully utilize the base class
event handlers.

Change-Id: I773e2eb939cbbf4bc9086cdf2b34e876597ea08e
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>