Don't crash when delivering stationary touchpoint as mouse release
QQuickPointerTouchEvent::touchEventForItem() can return a nullptr
so we have to check for that wherever we use it. In this crash scenario
while doing some kind of touch gesture it gets delivered as a synth
mouse event to a custom item.  When you release the last finger (or
release multiple fingers simultaneously), touchEventForItem() returns
nullptr because the fingers were no longer moving before the release,
and QQuickWindowPrivate::deliverTouchAsMouse() didn't check for that.

Task-number: QTBUG-56090
Change-Id: I7ada1f82711296722226fdc06d9c5add2422335b
Reviewed-by: Paolo Angelelli <paolo.angelelli@theqtcompany.com>
Reviewed-by: Jan Arve Sæther <jan-arve.saether@theqtcompany.com>