Move resizeOverlay() to QQuickPopupPrivate
This is a preparation step for making drawers movable. Drawer must
be later able to resize the overlay to fit the area of the window
that is covered by the drawer. For example, if the Drawer is below
the header, it can resize the overlay so that it won't be on top of
the header.

Change-Id: I2cfd025a31f3a517575f3dbf9b972dcd6957715c
Task-number: QTBUG-53168
Task-number: QTBUG-55360
Task-number: QTBUG-53609
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>