QQmlTypeLoader: Clean up trimCache()
Any type that:

a, has exactly one reference (from the type cache)
b, has finished loading, and is thus either in Error or Complete state
c, either has no compiled data or again exactly one reference on the
   compiled data (from the type)

is not needed anymore and can be trimmed.

The previous logic would not trim types with errors that have received
compiled data by loading from a file and it was generally harder to
understand.

Change-Id: Ieb9ce29599411ea5516742b874cbcf3dcab03bde
Reviewed-by: Michael Brasser <michael.brasser@live.com>
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>