Drawer: reparent to the overlay by default
Drawer is a special type of popup that always resides at one of the
window edges. Therefore it makes sense for drawers to operate on window
coordinates. Re-parenting drawers to the window overlay achieves that.

Now that the window overlay is guaranteed to always exist and the
ApplicationWindow.overlay attached property works even with plain QML
Window, we can reliably use it as a parent for Drawer.

Task-number: QTBUG-53168
Change-Id: I37c727001350217ea1d2d9c52d73b4cae44d7c8d
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>