Revert "Popup: expose flip API"
This reverts commit 09706e8f9fc481d848a3616cace58baf5bc8b67c.

It was a bit rushed to expose the allow*Flip properties. We have now
added several other similar internal allowFooBar flags, so clearly
a more sustainable solution is to craft some kind of horizontal and
vertical sizing/positioning policies out of them. So, remove this
unreleased API while we still can, so we don't have to deprecate it
right away.

Conflicts:
	src/imports/templates/qtquicktemplates2plugin.cpp

Change-Id: Iea33c11805071499b472b18426bbdc8d871a4a58
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>