Fix a crash in StackView::pop()
If an item is still activating (from a previous pop) when it gets
already popped out, we must not set the deactivating status before
calling prepareTransition(). This method cancels and finishes the
ongoing activation transition, and if the status says that the item
was deactivating, the item gets destroyed in the middle of preparing
for the deactivation transition.

Let prepareTransition() cancel any ongoing transition first, and then
set the status after the preparation. The cleanest way is to pass
the target status to startTransition() and completeTransition().

Task-number: QTBUG-56158
Change-Id: Id52752200b650ea9f84659bbf43431f8a8b22f1e
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>