Gallery: allow navigating back to the home page
Instead of replacing the home page and throwing it away for good, push
control specific introduction pages on top. Change the drawer button to
a back button to allow navigating back. Also, disable the drawer while
not on the home page to keep the navigation simple.

Thanks to Wolfgang Schneider for bringing up the idea in QTBUG-54300.

Change-Id: Ie0501332e27d7edab59eef32e59a27973b59ebf3
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>