Adding QQuickDesignerSupportMetaInfo::registerMockupObject()
This allows Qt Quick Designer to register a component with a custom parser
that can serve as a mockup for known C++ components registered
in e.g. main.cpp.

In many cases those components are the interface to the C++ backend.
While the C++ components itself are not relevant for the gui designer,
the user has to be able to instantiate gui components that use such C++
components.

Therefore we create mockup components with a custom parser, so all
property assignments are accepted.

Change-Id: I9c25752b0a5f11c2613346ffbbfd10e3b6eb3a6f
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>