Don't execute disabled transitions
If an enter or exit transition for a Drawer is disabled, don't
execute it at all when the drawer gets opened or closed. This
besides being more semantically correct, makes possible to control
manually the opening of a Drawer, for instance to implement an
external handle with which is possible to drag the drawer open or close.
If a Transition is disabled, is up to the caller to set correctly the
position value.

Change-Id: I5a0d96e3b7c948725dff2068f9d0a7ab1d3847e9
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>