Material: switch back to triggering ripples on press
There is something wrong with the ripple effect's speed calculation
when it's triggered on release. The ripple is supposed to grow slowly
on press, and fast on release. But when triggered on release, it grows
clearly too slow.

We also got a bug report complaining about the trigger on release
behavior. Ripples are triggered on press on Android 5, which is the
Android version with the most market share. The proposal is to fix
the ripple effect speed calculation when triggered on release later,
and switch to trigger on release when later versions of Android have
passed Android 5's market share.

Task-number: QTBUG-55520
Change-Id: I67994e0df3c318af24f0d1f843bb2d0c26c4d656
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>