Gallery: make the drawer usable with keyboard
Give it focus, and remove the check for the current index. It's a
leftover from the days when Gallery used to replace the current page
instead of doing a push. It looked silly to replace a page with the
same page. This can no longer happen, because the drawer is not even
accessible when navigated to a control page. And when navigating the
drawer with keyboard, it is exactly the current item that must be
triggered when you hit the spacebar. :)

Change-Id: I00b48fbe4dd61b06041cb0c3370dfa9c12505dcf
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>