TabButton: use Text's implicit size instead of content size
Using content size was a workaround (applied in 5f5654f) that helped
with mis-aligned tab labels, but broke the implicit size calculation
of TabButton itself. Now that the mis-alignment issue in QQuickText
has been fixed in qtdeclarative commit 79cfc87, we can restore the
correct implicit size calculation. This makes it possible to create
tabs that are not squeezed to fit in TabBar:

   TabBar {
       TabButton { text: "Foo"; width: implicitWidth }
       TabButton { text: "Bar"; width: implicitWidth }
       TabButton { text: "Baz"; width: implicitWidth }
   }

Task-number: QTBUG-55129
Change-Id: If0dbb00794299e324b7a8d8ce1370fe0a3491fe8
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>