Improve tst_tumbler
- Don't create a tumbler indiscriminately in init(). Some tests don't
use the tumbler, and it just creates more headaches for debugging, as
there is extra, unrelated qDebug output for the unused tumbler.
- Don't create the tumbler with createQmlObject(). Having code in a
string is difficult to maintain and doesn't have the benefit of e.g.
the static error checking that Creator can do.
- Instead of only automatically cleaning up the tumbler item in
cleanup(), introduce a "cleanupItem" that all dynamically created items
can use as their parent so that they too can benefit from automatic
cleanup.
- Comment out incomplete test and mention that it's a TODO.

Change-Id: I070fbb018bc9270555774bb0de2d72b7889edb35
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>