Flickable: respect keepMouseGrab again
The keep grab property works in a "cooperative" way, items need to
respect it. Since we don't pre-grab any more, flickable would check the
wrong item for the property. It should simply take the receiver item
into account, which has to have the grab if anyone has it. If the
receiver didn't have the grab, it wouldn't be the receiver in the first
place.

Task-number: QTBUG-55229
Change-Id: If7d8b0c99bf912c0e09427cd20e89be1798ace5d
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>