add qt.quick.pointer.grab logging category to track pointer grab
This is the ultimate destination of all grab/ungrab functions, both the
legacy ones in QQuickWindow and the potential replacements.  Since we
rearchitected event delivery significantly in 5.8, it will be important
to debug any behavior change issues that arise in the field; and in many
such cases it's important to track which eventpoint on which device
is being grabbed by which item.

It generates output like
[     2.012 D] QQuickEventPoint::setGrabber - "Advanced Silicon" point 1c1 Updated : grab QQuickItem(0) -> MapComponent_QMLTYPE_7(0x25d0900, parent=0x23cb430, geometry=0,0 360x620)
[     2.012 D] QQuickEventPoint::setGrabber - "core pointer    " point 0 Updated : grab QQuickMouseArea_QML_10(0x25a3dd0, name="big MA", parent=0x25d0900, geometry=0,0 360x620) -> QQuickItem(0)

Change-Id: Ic98c3c91c719b430590530ba8c5a7eff8063d330
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@qt.io>