Fix QQuickToolTip's delay and visibility
The imperative open() and close() methods used to be the only ways to
open and close popups, but the visible-property was later made writable
to allow declarative visibility bindings. Since then, it is no longer
sufficient for QQuickToolTip to overshadow open() and close() in QML,
because setting the visible-property would bypass these overshadowed
method.

There was a bit of duplicate code between setVisible(), open(), and
close(). This change moves the logic to one place by changing open()
and close() to call setVisible(). Furthermore, setVisible() has been
made virtual to make it possible for QQuickToolTip to apply its delay
properly. QQuickToolTip needs to control the delay and timeout timers
before the effective visibility is applied on the popup.

Task-number: QTBUG-55572
Change-Id: I5a109157f9ec5d0db145e710426665a9a8d7e870
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>