Make a plain AbstractButton calculate suitable implicit size
Just like commit 87b21a57 made Control do it, this applies the same
logic to AbstractButton. The simple reasoning is that when customizing
both the contentItem and the background, it is much better to use the
AbstractButton type that does not have any default building blocks.
Previously you would have to supply implicit size calculation as well,
but this makes it much more convenient to use.

[ChangeLog][Controls][AbstractButton] A plain AbstractButton now
calculates its implicit size based on the implicit size of the content
item plus paddings, and the implicit size of the background item.

Change-Id: Ic380366b25940fbb9f28882c4622d58f4c042a07
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>