Fix QQuickControl::setFocusReason()
There is a little behavior change in focus handling in Qt Quick after
qdeclarative commit e7da97b. The order of QQuickItem::focusOutEvent()
and QQuickItem::ItemActiveFocusHasChanged have changed. Now we need to
emit visualFocusChanged() in setFocusReason() when losing visual focus
to ensure that bindigs to visualFocus get re-evaluated as appropriate.

Change-Id: I52e728c7df751a54b36f353415b2666cfedc73ad
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>