Fix release event delivery in deliverTouchAsMouse
Originally in the complicated version of this part of the code, the
event would always end up in the right place, since the grabber was kept
and events delivered reliably. Now we take device and point ID for the
grabs into account. When delivering a mouse event from a touch, we would
clear the pointer device and ID before trying to send the release, which
resulted in the release event never being delivered.

Change-Id: I9bc1b2a795023740f8fe3961d779bc8bff0dc35e
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>