QQuickPopupPrivate: unbreak ubsan build
The tst_drawer test uses members of the class, but it was not (autotest-)
exported. That's no problem as long as the function called is inline, or
a data member is accessed, but in a UBSan build, the typeid of a poly-
morphic class is checked on each access, so the typeinfo object must be
exported.

Fix by autotest-exporting the class and defining the dtor out-of-line to
pin the typeinfo and vtabe to a single TU.

Change-Id: I1524e63a330dbdb0ae1ceb60263b589734bfd3aa
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>