SwipeDelegate: don't emit clicked when released outside
If there are no delegates and hence the mouse hasn't been grabbed, we
should clear the pressed state when the mouse is dragged outside the
control. We can do so by falling back to the event handling of the
base class (QQuickItemDelegate) when we have no delegates.
This also ensures that the canceled() signal is emitted.

A similar thing is done for controls with delegates, except that only
the vertical position of the mouse is checked, as we still want to
initiate swipes horizontally.

Change-Id: I7738f5b9e8e8b6ce4a733008fa4ff73596e854ea
Task-number: QTBUG-56312
Task-number: QTBUG-57285
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>