Remove wrong Container revision registration
We register Container.qml for QtQuick.Controls 2.0, so it is important
to not override the registration with plain QQuickContainer template
revision 1 for QtQuick.Controls 2.1, or else the behavior of Container
will be different depending on whether the user imports QtQuick
Controls 2.0 or 2.1. The auto test started failing when the import
version was bumped from 2.0 to 2.1. This revealed that "Container" was
pointing to Container.qml in 2.0, and to QQuickContainer in 2.1.

Change-Id: I80b8de3bf556c57e4ae853b116d387431dfa02c7
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>