QQuickCanvasItem: use QStringLiteral less
QStringLiteral was used to store type of image, but
it was needless wrapper, because it always was be converted
to char* to pass to QImage::save(), so replace it with char*
to save allocations and text size.

Also replace QStringLiteral().arg().arg() pattern with
QStringBuilder. Again save allocations and text size.

Change-Id: I6614e127789a05069da200c263145654a289b309
Reviewed-by: Gunnar Sletta <gunnar@sletta.org>