Gallery: fix the height of the settings dialog
The settings dialog used to be a plain Popup, which does not provide
buttons. There were manually created Ok and Cancel buttons in a column
together with the other contents, and the height of the column was used
to calculate the total height of the dialog. Now that the new Dialog
type with support for standard buttons is used, the height of the
column is no longer sufficient since the dialog buttons are not part of
it. Just let the dialog calculate the appropriate height based on its
width and content.

Change-Id: Ib4edf6e525214429887ad4ec6371ac0b5624c2e9
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>